Update derivative1.tex
For derivatives we add one to the index of the sum. So here since the original f starts at k=0 then the second derivative has to start from k=2. I corrected the two f double prime sums to start at 2. Also in the ratio test method (method 1) it is more consistent to have the top x to be to power 2(n+1)+1 and bottom x to be to power 2n+1. You get the same answer as what was already there but this is more consistent when applying the ratio test.