Update inversesTrig7.tex
https://ximera.osu.edu/mooculus/derivativesOfInverseFunctions/exercises/exerciseList/derivativesOfInverseFunctions/exercises/inversesTrig7

Changed all tan^(-1) to arctan because the answers have to be in arctan so it is confusing to ask one way and want answers in another way. Also took out a step in finding the derivative because it follows from the line above and no need to first write it as d/dx of the function.